# openapi.yaml
openapi: 3.1.0
info:
  title: Trustify
  description: Software Supply-Chain Security API
  license:
    name: Apache License, Version 2.0
    identifier: Apache-2.0
  version: 0.5.0-beta.2
paths:
...
  /api/v3/purl:
    get:
      tags:
      - purl
      summary: List fully-qualified pURLs
      operationId: listPurl
      parameters:
      - name: q
        in: query
        description: |
          EBNF grammar for the _q_ parameter:
          ```text
              q = ( values | filter ) { '&' q }
              values = value { '|', values }
              filter = field, operator, values
              operator = "=" | "!=" | "~" | "!~" | ">=" | ">" | "<=" | "<"
              value = (* any text but escape special characters with '\' *)
              field = (* must match an entity attribute name *)
          ```
          Any values in a _q_ will result in a case-insensitive "full
          text search", effectively producing an OR clause of LIKE
          clauses for every string-ish field in the resource being
          queried.

          Examples:
          - `foo` - any field containing 'foo'
          - `foo|bar` - any field containing either 'foo' OR 'bar'
          - `foo&bar` - some field contains 'foo' AND some field contains 'bar'

          A _filter_ may also be used to constrain the results. The
          filter's field name must correspond to one of the resource's
          attributes. If it doesn't, an error will be returned
          containing a list of the valid fields for that resource.

          An ASCII value of `NUL`, percent-encoded as `%00`, may be used
          to find resources on which a particular field isn't set. For
          example, `name=%00` and `name!=%00` yield the WHERE clauses,
          'NAME IS NULL' and 'NAME IS NOT NULL', respectively.

          Examples:
          - `name=foo` - entity's _name_ matches 'foo' exactly
          - `name~foo` - entity's _name_ contains 'foo', case-insensitive
          - `name~foo|bar` - entity's _name_ contains either 'foo' OR 'bar', case-insensitive
          - `name=` - entity's _name_ is the empty string, ''
          - `name=%00` - entity's _name_ isn't set
          - `published>3 days ago` - date values can be "human time"

          Multiple full text searches and/or filters should be
          '&'-delimited -- they are logically AND'd together.

          - `red hat|fedora&labels:type=cve|osv&published>last wednesday 17:00`

          Fields corresponding to JSON objects in the database may use a
          ':' to delimit the column name and the object key,
          e.g. `purl:qualifiers:type=pom`

          Any operator or special character, e.g. '|', '&', within a
          value should be escaped by prefixing it with a backslash.
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: |
          EBNF grammar for the _sort_ parameter:
          ```text
              sort = field [ ':', order ] { ',' sort }
              order = ( "asc" | "desc" )
              field = (* must match the name of entity's attributes *)
          ```
          The optional _order_ should be one of "asc" or "desc". If
          omitted, the order defaults to "asc".

          Each _field_ name must correspond to one of the columns of the
          table holding the entities being queried. Those corresponding
          to JSON objects in the database may use a ':' to delimit the
          column name and the object key,
          e.g. `purl:qualifiers:type:desc`
        required: false
        schema:
          type: string
      - name: offset
        in: query
        description: |-
          The first item to return, skipping all that come before it.

          NOTE: The order of items is defined by the API being called.
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: limit
        in: query
        description: |-
          The maximum number of entries to return.

          Zero means: return no items (the total count is still computed if requested).
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: total
        in: query
        description: Whether to compute and return the total count of matching items.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: All relevant matching qualified PURLs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResults_PurlSummary'
...
  /api/v3/purl/recommend:
    post:
      tags:
      - purl
      operationId: recommend
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendRequest'
        required: true
      responses:
        '200':
          description: Get recommendations and remediations for provided purls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendResponse'
  /api/v3/purl/{key}:
    get:
      tags:
      - purl
      summary: Retrieve details of a fully-qualified pURL
      operationId: getPurl
      parameters:
      - name: deprecated
        in: query
        required: false
        schema:
          type: string
          enum:
          - Ignore
          - Consider
      - name: key
        in: path
        description: opaque identifier for a fully-qualified PURL, or URL-encoded pURL itself
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Details for the qualified PURL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurlDetails'
...
components:
  schemas:
...
    AdvisoryHead:
      type: object
      required:
      - uuid
      - identifier
      - document_id
      - issuer
      - published
      - withdrawn
      - title
      - labels
      properties:
        document_id:
          type: string
          description: The identifier of the advisory, as provided by the document.
        identifier:
          type: string
          description: The identifier of the advisory, as assigned by the issuing organization.
        issuer:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/OrganizationSummary'
            description: |-
              The issuer of the advisory, if known. If no issuer is able to be
              determined, this field will not be included in a response.
        labels:
          $ref: '#/components/schemas/Labels'
          description: Informational labels attached by the system or users to this advisory.
        modified:
          type:
          - string
          - 'null'
          format: date-time
          description: The date (in RFC3339 format) of when the advisory was last modified, if any.
        published:
          type:
          - string
          - 'null'
          format: date-time
          description: The date (in RFC3339 format) of when the advisory was published, if any.
        title:
          type:
          - string
          - 'null'
          description: The title of the advisory as assigned by the issuing organization.
        uuid:
          type: string
          description: The opaque UUID of the advisory.
        withdrawn:
          type:
          - string
          - 'null'
          format: date-time
          description: The date (in RFC3339 format) of when the advisory was withdrawn, if any.
...
    BasePurlHead:
      type: object
      required:
      - uuid
      - purl
      properties:
        purl:
          $ref: '#/components/schemas/Purl'
          description: The actual base PURL
        uuid:
          type: string
          format: uuid
          description: The ID of the base PURL
...
    BaseScore:
      type: object
      description: |-
        Base score information in the context of a [`VulnerabilityHead`]. Notably, this excludes the
        raw CVSS vector string.

        Uses the same `ScoreType` and `Severity` serialization as [`crate::common::model::Score`]
        so that all score-related fields in API responses are consistent.
      required:
      - type
      - severity
      - score
      properties:
        score:
          type: number
          format: double
        severity:
          $ref: '#/components/schemas/Severity'
        type:
          $ref: '#/components/schemas/ScoreType'
...
    Labels:
      type: object
      additionalProperties:
        type: string
    LicenseCategory:
      type: string
      enum:
      - declared
      - concluded
    LicenseInfo:
      type: object
      required:
      - license_name
      - license_type
      properties:
        license_name:
          type: string
        license_type:
          $ref: '#/components/schemas/LicenseCategory'
    LicenseRefMapping:
      type: object
      required:
      - license_id
      - license_name
      properties:
        license_id:
          type: string
        license_name:
          type: string
...
    OrganizationHead:
      type: object
      description: |-
        An organization who may issue advisories, product SBOMs, or
        otherwise be involved in supply-chain evidence.
      required:
      - id
      - name
      - cpe_key
      - website
      properties:
        cpe_key:
          type:
          - string
          - 'null'
          description: The `CPE` key of the organization, if known.
        id:
          type: string
          format: uuid
          description: The opaque UUID of the organization.
        name:
          type: string
          description: The name of the organization.
        website:
          type:
          - string
          - 'null'
          description: The website of the organization, if known.
    OrganizationSummary:
      allOf:
      - $ref: '#/components/schemas/OrganizationHead'
...
    PaginatedResults_PurlSummary:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            allOf:
            - $ref: '#/components/schemas/PurlHead'
            - type: object
              required:
              - base
              - version
              - qualifiers
              properties:
                base:
                  $ref: '#/components/schemas/BasePurlHead'
                qualifiers:
                  type: object
                  additionalProperties:
                    type: string
                  propertyNames:
                    type: string
                  deprecated: true
                version:
                  $ref: '#/components/schemas/VersionedPurlHead'
        total:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
...
    Purl:
      type: string
      format: uri
    PurlAdvisory:
      allOf:
      - $ref: '#/components/schemas/AdvisoryHead'
      - type: object
        required:
        - status
        properties:
          status:
            type: array
            items:
              $ref: '#/components/schemas/PurlStatus'
    PurlDetails:
      allOf:
      - $ref: '#/components/schemas/PurlHead'
      - type: object
        required:
        - version
        - base
        - advisories
        - licenses
        - licenses_ref_mapping
        properties:
          advisories:
            type: array
            items:
              $ref: '#/components/schemas/PurlAdvisory'
          base:
            $ref: '#/components/schemas/BasePurlHead'
          licenses:
            type: array
            items:
              $ref: '#/components/schemas/LicenseInfo'
          licenses_ref_mapping:
            type: array
            items:
              $ref: '#/components/schemas/LicenseRefMapping'
            deprecated: true
          version:
            $ref: '#/components/schemas/VersionedPurlHead'
    PurlHead:
      type: object
      required:
      - uuid
      - purl
      properties:
        purl:
          $ref: '#/components/schemas/Purl'
          description: The actual qualified PURL
        uuid:
          type: string
          format: uuid
          description: The ID of the qualified PURL
    PurlStatus:
      type: object
      required:
      - vulnerability
      - advisory
      - scores
      - status
      - context
      properties:
        advisory:
          $ref: '#/components/schemas/AdvisoryHead'
        context:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/StatusContext'
        scores:
          type: array
          items:
            $ref: '#/components/schemas/ScoredVector'
          description: All CVSS scores associated with the vulnerability
        status:
          type: string
        version_range:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/VersionRange'
        vulnerability:
          $ref: '#/components/schemas/VulnerabilityHead'
...
    RecommendEntry:
      type: object
      required:
      - package
      - vulnerabilities
      properties:
        package:
          type: string
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/VulnerabilityStatus'
    RecommendRequest:
      type: object
      required:
      - purls
      properties:
        purls:
          type: array
          items:
            $ref: '#/components/schemas/Purl'
    RecommendResponse:
      type: object
      required:
      - recommendations
      properties:
        recommendations:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/RecommendEntry'
          propertyNames:
            type: string
...
    RemediationCategory:
      type: string
      enum:
      - vendor_fix
      - workaround
      - mitigation
      - no_fix_planned
      - none_available
      - will_not_fix
    RemediationSummary:
      type: object
      required:
      - category
      - data
      properties:
        category:
          $ref: '#/components/schemas/RemediationCategory'
        data:
          description: For internal use only. May be removed at any point and should not be used.
          deprecated: true
        details:
          type:
          - string
          - 'null'
        url:
          type:
          - string
          - 'null'
...
    Score:
      type: object
      description: 'A parsed CVSS score: the scoring system version, numeric value, and derived severity.'
      required:
      - type
      - value
      - severity
      properties:
        severity:
          $ref: '#/components/schemas/Severity'
          description: The severity band derived from the score value.
        type:
          $ref: '#/components/schemas/ScoreType'
          description: The scoring system version.
        value:
          type: number
          format: double
          description: The numeric score, rounded to one decimal place.
      example:
        type: '3.1'
        value: 7.5
        severity: high
    ScoreType:
      type: string
      description: The type of score, indicating the scoring system and version used.
      enum:
      - '2.0'
      - '3.0'
      - '3.1'
      - '4.0'
      x-enum-descriptions:
      - CVSS v2.0 score
      - CVSS v3.0 score
      - CVSS v3.1 score
      - CVSS v4.0 score
    ScoredVector:
      allOf:
      - $ref: '#/components/schemas/Score'
        description: The score type, value, and derived severity.
      - type: object
        required:
        - vector
        properties:
          vector:
            type: string
            description: The raw CVSS vector string (e.g. `CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H`).
      description: |-
        A CVSS score combined with its raw vector string, for contexts where clients need both
        the pre-parsed numeric values and the original vector for display or re-parsing.
      example:
        type: '3.1'
        value: 7.5
        severity: high
        vector: CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H
    Severity:
      type: string
      description: Severity rating derived from a CVSS score value.
      enum:
      - none
      - low
      - medium
      - high
      - critical
      x-enum-descriptions:
      - No impact (score = 0.0)
      - Low severity (score 0.1–3.9)
      - Medium severity (score 4.0–6.9)
      - High severity (score 7.0–8.9)
      - Critical severity (score 9.0–10.0)
...
    StatusContext:
      oneOf:
      - type: object
        required:
        - purl
        properties:
          purl:
            $ref: '#/components/schemas/Purl'
      - type: object
        required:
        - cpe
        properties:
          cpe:
            type: string
...
    VersionRange:
      oneOf:
      - type: object
        required:
        - version_scheme_id
        - low_version
        - low_inclusive
        - high_version
        - high_inclusive
        properties:
          high_inclusive:
            type: boolean
          high_version:
            type: string
          low_inclusive:
            type: boolean
          low_version:
            type: string
          version_scheme_id:
            type: string
      - type: object
        required:
        - version_scheme_id
        - low_version
        - low_inclusive
        properties:
          low_inclusive:
            type: boolean
          low_version:
            type: string
          version_scheme_id:
            type: string
      - type: object
        required:
        - version_scheme_id
        - high_version
        - high_inclusive
        properties:
          high_inclusive:
            type: boolean
          high_version:
            type: string
          version_scheme_id:
            type: string
      - type: 'null'
        default: null
    VersionedPurlHead:
      type: object
      required:
      - uuid
      - purl
      - version
      properties:
        purl:
          $ref: '#/components/schemas/Purl'
          description: The actual, versioned PURL
        uuid:
          type: string
          format: uuid
          description: The ID of the versioned PURL
        version:
          type: string
          description: The version from the PURL
...
    VexJustification:
      oneOf:
      - type: string
        enum:
        - ComponentNotPresent
      - type: string
        enum:
        - VulnerableCodeNotPresent
      - type: string
        enum:
        - VulnerableCodeNotInExecutePath
      - type: string
        enum:
        - VulnerableCodeCannotBeControlledByAdversary
      - type: string
        enum:
        - InlineMitigationsAlreadyExist
      - type: string
        enum:
        - NotProvided
      - type: object
        required:
        - Other
        properties:
          Other:
            type: string
    VexStatus:
      oneOf:
      - type: string
        enum:
        - Affected
      - type: string
        enum:
        - Fixed
      - type: string
        enum:
        - NotAffected
      - type: string
        enum:
        - UnderInvestigation
      - type: string
        enum:
        - Recommended
      - type: object
        required:
        - Other
        properties:
          Other:
            type: string
...
    VulnerabilityHead:
      type: object
      required:
      - normative
      - identifier
      - title
      - description
      - reserved
      - published
      - modified
      - withdrawn
      - discovered
      - released
      - cwes
      properties:
        base_score:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/BaseScore'
            description: The main, base score.
        cwes:
          type: array
          items:
            type: string
          description: Associated CWE, if any.
        description:
          type:
          - string
          - 'null'
          description: The description of the vulnerability, if known.
        discovered:
          type:
          - string
          - 'null'
          format: date-time
          description: The date (in RFC3339 format) of when the vulnerability was discovered, if any.
        identifier:
          type: string
          description: |-
            The globally-unique identifier for the vulnerability.
            Traditionally (but not required) refers to the assigned
            CVE identifier.
        modified:
          type:
          - string
          - 'null'
          format: date-time
          description: The date (in RFC3339 format) of when the vulnerability was last modified, if any.
        normative:
          type: boolean
        published:
          type:
          - string
          - 'null'
          format: date-time
          description: The date (in RFC3339 format) of when the vulnerability was published, if any.
        released:
          type:
          - string
          - 'null'
          format: date-time
          description: The date (in RFC3339 format) of when software containing the vulnerability first released, if known.
        reserved:
          type:
          - string
          - 'null'
          format: date-time
          description: The date (in RFC3339 format) of when the vulnerability identifier was reserved, if any.
        title:
          type:
          - string
          - 'null'
          description: The title of the vulnerability, if known.
        withdrawn:
          type:
          - string
          - 'null'
          format: date-time
          description: The date (in RFC3339 format) of when the vulnerability was last withdrawn, if any.
...
    VulnerabilityStatus:
      type: object
      required:
      - id
      - remediations
      properties:
        id:
          type: string
        justification:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/VexJustification'
        remediations:
          type: array
          items:
            $ref: '#/components/schemas/RemediationSummary'
        status:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/VexStatus'
...